+++ pkg/dateparser/dateparser_test.go
@@ -18,6 +18,7 @@
 package dateparser
 
 import (
+	"encoding/json"
 	"testing"
 	"time"
 
@@ -90,4 +91,22 @@ func TestFactoryShouldCreateAMask(t *testing.T) {
 	assert.Equal(t, mask, config, "should be equal")
 	assert.True(t, present, "should be true")
 	assert.Nil(t, err, "error should be nil")
+}
+
+func TestMaskingShouldReplaceUnixEpochByDateString(t *testing.T) {
+	outputFormat := "02/01/06"
+	dateMask := NewMask("unixEpoch", outputFormat)
+	data := json.Number("1647512434")
+	resulttime, err := dateMask.Mask(data)
+	assert.Equal(t, nil, err, "error should be nil")
+	assert.Equal(t, "17/03/22", resulttime, "Should return the same time")
+}
+
+func TestMaskingShouldReplaceDateStringByUnixEpoch(t *testing.T) {
+	inputFormat := "02/01/06"
+	dateMask := NewMask(inputFormat, "unixEpoch")
+	data := "17/03/22"
+	result, err := dateMask.Mask(data)
+	assert.Equal(t, nil, err, "error should be nil")
+	assert.Equal(t, int64(1647475200), result, "Should return the same time")
 }